Preparations for v5.38 "file" command compatibility update

+++ src/main/java/com/j256/simplemagic/types/BigEndianString16Type.java
@@ -1,5 +1,7 @@
 package com.j256.simplemagic.types;
 
+import java.math.BigInteger;
+
 /**
  * A two-byte unicode (UCS16) string in big-endian byte order.
  * 
@@ -24,8 +26,8 @@ public class BigEndianString16Type extends StringType {
 	}
 
 	@Override
-	public Object isMatch(Object testValue, Long andValue, boolean unsignedType, Object extractedValue,
-			MutableOffset mutableOffset, byte[] bytes) {
+	public Object isMatch(Object testValue, BigInteger andValue, boolean unsignedType, Object extractedValue,
+						  MutableOffset mutableOffset, byte[] bytes) {
 		// we do the match on the extracted chars
 		char[] chars = (char[]) extractedValue;
 		return super.findOffsetMatch((TestInfo) testValue, mutableOffset.offset, mutableOffset, null, chars,
